refactor(voices): extract the studio view model out of the controller
The Voices controller carried a DOM-free half — host scope (VOICE_HOSTS,
resolveInitialHost, LAST_HOST_KEY), the derived StudioViewModel, and the
#474 twin-name strategy — inside a class whose other job is painting. It
was always pure, but living in the controller meant every assertion about
pin resolution, menu seating or twin-name strategy had to be made through
rendered DOM.

Moved verbatim to voices-view-model.ts (design §14). `viewModel` becomes a
free function; the controller re-exports VoiceHostId so its public surface
is unchanged. No behaviour change — the existing 53 controller tests pass
untouched.
